Split libolaserver into two parts to stop dpkg-shlibdeps warnings
