+++ olad/Makefile.mk
@@ -6,28 +6,13 @@ dist_noinst_DATA += olad/testdata/test_preferences.conf
 ##################################################
 oladinclude_HEADERS += olad/OlaDaemon.h olad/OlaServer.h
 
-noinst_HEADERS += olad/TestCommon.h
-
 # LIBRARIES
 ##################################################
-ola_server_plugin_interface_sources = \
-    olad/Device.cpp \
-    olad/DeviceManager.cpp \
-    olad/DeviceManager.h \
-    olad/Plugin.cpp \
-    olad/PluginAdaptor.cpp \
-    olad/PortBroker.cpp \
-    olad/PortManager.cpp \
-    olad/PortManager.h
-
 ola_server_sources = \
-    olad/Client.cpp \
-    olad/Client.h \
     olad/ClientBroker.cpp \
     olad/ClientBroker.h \
     olad/DiscoveryAgent.cpp \
     olad/DiscoveryAgent.h \
-    olad/DmxSource.cpp \
     olad/DynamicPluginLoader.cpp \
     olad/DynamicPluginLoader.h \
     olad/HttpServerActions.h \
@@ -37,12 +22,7 @@ ola_server_sources = \
     olad/PluginLoader.h \
     olad/PluginManager.cpp \
     olad/PluginManager.h \
-    olad/Port.cpp \
-    olad/Preferences.cpp \
-    olad/RDMHTTPModule.h \
-    olad/Universe.cpp \
-    olad/UniverseStore.cpp \
-    olad/UniverseStore.h
+    olad/RDMHTTPModule.h
 ola_server_additional_libs =
 
 if HAVE_DNSSD
@@ -63,14 +43,7 @@ ola_server_additional_libs += common/http/libolahttp.la
 endif
 
 # lib olaserver
-lib_LTLIBRARIES += olad/libolaserverplugininterface.la olad/libolaserver.la
-
-olad_libolaserverplugininterface_la_SOURCES = \
-    $(ola_server_plugin_interface_sources)
-olad_libolaserverplugininterface_la_CXXFLAGS = $(COMMON_CXXFLAGS)
-olad_libolaserverplugininterface_la_LIBADD = common/libolacommon.la \
-                                             common/web/libolaweb.la \
-                                             ola/libola.la
+lib_LTLIBRARIES += olad/plugin_api/libolaserverplugininterface.la olad/libolaserver.la
 
 olad_libolaserver_la_SOURCES = $(ola_server_sources) \
                                olad/OlaServer.cpp \
@@ -81,7 +54,7 @@ olad_libolaserver_la_LIBADD = $(PLUGIN_LIBS) \
                               common/libolacommon.la \
                               common/web/libolaweb.la \
                               ola/libola.la \
-                              olad/libolaserverplugininterface.la \
+                              olad/plugin_api/libolaserverplugininterface.la \
                               $(ola_server_additional_libs)
 # Simon: I'm not too sure about this but it seems that because PLUGIN_LIBS is
 # determined at configure time, we need to add them here.
@@ -106,39 +79,17 @@ endif
 # TESTS
 ##################################################
 test_programs += \
-    olad/ClientTester \
-    olad/DeviceTester \
-    olad/OlaTester \
-    olad/PortTester \
-    olad/UniverseTester
+    olad/OlaTester
 
 COMMON_OLAD_TEST_LDADD = $(COMMON_TESTING_LIBS) $(libprotobuf_LIBS) \
+                         olad/plugin_api/libolaserverplugininterface.la \
                          olad/libolaserver.la \
-                         olad/libolaserverplugininterface.la \
                          common/libolacommon.la
 
-olad_ClientTester_SOURCES = olad/ClientTest.cpp
-olad_ClientTester_CXXFLAGS = $(COMMON_TESTING_FLAGS)
-olad_ClientTester_LDADD = $(COMMON_OLAD_TEST_LDADD)
-
-olad_DeviceTester_SOURCES = olad/DeviceTest.cpp olad/DeviceManagerTest.cpp
-olad_DeviceTester_CXXFLAGS = $(COMMON_TESTING_FLAGS)
-olad_DeviceTester_LDADD = $(COMMON_OLAD_TEST_LDADD)
-
-olad_PortTester_SOURCES = olad/PortTest.cpp olad/PortManagerTest.cpp
-olad_PortTester_CXXFLAGS = $(COMMON_TESTING_FLAGS)
-olad_PortTester_LDADD = $(COMMON_OLAD_TEST_LDADD)
-
 olad_OlaTester_SOURCES = \
-    olad/DmxSourceTest.cpp \
     olad/PluginManagerTest.cpp \
-    olad/PreferencesTest.cpp \
     olad/OlaServerServiceImplTest.cpp
 olad_OlaTester_CXXFLAGS = $(COMMON_TESTING_FLAGS)
 olad_OlaTester_LDADD = $(COMMON_OLAD_TEST_LDADD)
-
-olad_UniverseTester_SOURCES = olad/UniverseTest.cpp
-olad_UniverseTester_CXXFLAGS = $(COMMON_TESTING_FLAGS)
-olad_UniverseTester_LDADD = $(COMMON_OLAD_TEST_LDADD)
 
 CLEANFILES += olad/ola-output.conf
