+++ olad/OlaServer.cpp
@@ -43,9 +43,7 @@
 #include "ola/rdm/PidStore.h"
 #include "ola/rdm/UID.h"
 #include "ola/stl/STLUtils.h"
-#include "olad/Client.h"
 #include "olad/ClientBroker.h"
-#include "olad/DeviceManager.h"
 #include "olad/DiscoveryAgent.h"
 #include "olad/OlaServer.h"
 #include "olad/OlaServerServiceImpl.h"
@@ -54,10 +52,12 @@
 #include "olad/PluginManager.h"
 #include "olad/Port.h"
 #include "olad/PortBroker.h"
-#include "olad/PortManager.h"
 #include "olad/Preferences.h"
 #include "olad/Universe.h"
-#include "olad/UniverseStore.h"
+#include "olad/plugin_api/Client.h"
+#include "olad/plugin_api/DeviceManager.h"
+#include "olad/plugin_api/PortManager.h"
+#include "olad/plugin_api/UniverseStore.h"
 
 #ifdef HAVE_LIBMICROHTTPD
 #include "olad/OladHTTPServer.h"
