+++ olad/PluginManagerTest.cpp
@@ -28,7 +28,7 @@
 #include "olad/PluginLoader.h"
 #include "olad/PluginManager.h"
 #include "olad/Preferences.h"
-#include "olad/TestCommon.h"
+#include "olad/plugin_api/TestCommon.h"
 #include "ola/testing/TestUtils.h"
