+++ olad/plugin_api/Device.cpp
