+++ olad/plugin_api/DeviceManager.h
