+++ olad/plugin_api/Plugin.cpp
