+++ olad/plugin_api/UniverseTest.cpp
@@ -31,16 +31,16 @@
 #include "ola/rdm/RDMReply.h"
 #include "ola/rdm/RDMResponseCodes.h"
 #include "ola/rdm/UID.h"
-#include "olad/Client.h"
 #include "olad/DmxSource.h"
 #include "olad/PluginAdaptor.h"
 #include "olad/Port.h"
 #include "olad/PortBroker.h"
-#include "olad/PortManager.h"
 #include "olad/Preferences.h"
-#include "olad/TestCommon.h"
 #include "olad/Universe.h"
-#include "olad/UniverseStore.h"
+#include "olad/plugin_api/Client.h"
+#include "olad/plugin_api/PortManager.h"
+#include "olad/plugin_api/TestCommon.h"
+#include "olad/plugin_api/UniverseStore.h"
 #include "ola/testing/TestUtils.h"
